+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [unreleased]
 ### Fixed
 - Upload to Codecov step in `tests_n_coverage.yml` action
+- Fix `multipleConditionExplanation` to be lowercase
 
 ## [2.5.1]
 ### Fixed